+++ src/plugins/textReplace/style.css
@@ -1,0 +1,1 @@
+/* meow */
